唯一登录的代码实现

+++ src/main/java/com/laoxing/skill/entity/Goods.java
@@ -13,6 +13,6 @@ public class Goods {
     private Integer gid;
     private String gname;
     private Double gprice;
-    private Double gstock;
+    private int gstock;
     private String gimg;
 }
